+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   
 ### Fixed
 - [#61] Don't modify an instrumented function's parameters when rendering them.
+- Correct the structure of the `return_value` object in a `return` event.
+
 ## [0.5.0] - 2021-03-08
 ### Added
 - Packages in config file can now be set for 'shallow' tracking. This eliminates most of